[CP-3194] 'DeviceDetached' event logic moved to main device-manager domain (mds logic)
