+++ libs/core/__deprecated__/renderer/components/core/modal/modal-service-emitter.ts
@@ -1,0 +1,12 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+import { EventEmitter } from "events"
+
+export enum ModelEvent {
+  Close = "modal/close"
+}
+
+export const modalEventEmitter = new EventEmitter()
