+++ libs/core/overview/components/overview-screens/pure-overview/pure-overview.component.tsx
@@ -25,9 +25,8 @@ import { FunctionComponent } from "Core/core/types/function-component.interface"
 import React, { useCallback, useEffect, useState } from "react"
 import { CheckForUpdateState } from "Core/update/constants/check-for-update-state.constant"
 import { useWatchDeviceDataEffect } from "Core/overview/components/overview-screens/helpers/use-watch-device-data-effect"
-import { useDeactivateDeviceAndRedirect } from "Core/overview/components/overview-screens/pure-overview/use-deactivate-device-and-redirect.hook"
 import { useSelector } from "react-redux"
-import { isActiveDeviceAttachedSelector } from "device-manager/feature"
+import { isActiveDeviceAttachedSelector, useDeactivateDeviceAndRedirect } from "device-manager/feature"
 import { selectDeviceErrorModalOpened } from "generic-view/store"
 import { useHelpShortcut } from "help/store"
