+++ libs/device-manager/feature/src/hooks/index.ts
@@ -3,6 +3,7 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
+export * from "./use-deactivate-device-and-redirect.hook"
 export * from "./use-device-manager-connect-failed"
 export * from "./use-device-manager-connected"
 export * from "./use-device-manager-detached"
